ReduceNeighbourhood: support local (statistical) standardization.
'Standardization' is a statistical procedure used to transform a distribution so that the mean is zero and the standard
deviation is one. The implementation provided in this commit is a *locally adaptive* standardization procedure, so
within the local neighbourhood, each voxel value is standardized.

Note, however, that because the neighbourhood itself also changes, in general the neighbourhood will not be
statistically standardized after this routine. If the neighbourhood is taken to be extremely large then the whole image
volume will be statistically standardized. However, this routine is tuned for smaller neighbourhoods and will compute
large neighbourhoods with extreme inefficiency. A simpler implementation with an infinitely-large neighbourhood could be
easily and efficiently computed using the machinery of the NormalizePixels operation.