Surface_Meshes: pruned excess headers.
Also added a WIP script to locate irrelevant headers. It won't work correctly (all the time) when transitive includes
are possible, but should work OK with small manual touch-ups.

The compile script also now forwards the return value of the actual compilation, though I'm not sure it is needed since
the script *should* fail as soon as a compilation fails (set -e).