Expanded README.
This commit increases the amount of information displayed in the 'about' section. The idea is to showcase what exactly
DICOMautomaton supports since prior to this commit potential users were implcitly asked to find the documentation, scan
through the reference guide, and possibly find an example on their own -- just to get an idea of how the system is even
used. While there still is not an explicit example in the README, at least the functionality may be compelling enough
for users to investigate further.