CI: build an AppImage.
This commit does not yet result in published build artifacts, and the CI still builds on Arch Linux. In order to
minimize glibc issues in the AppImage, the CI pipeline will eventually have to switch to Debian. The extraction script
added in this commit can handle both Debian and Arch Linux.