CI: disable background memory notification process.
Previously this process could run in the background without impacting when the CI jobs terminate, but it not seems to
cause the build stages to hang indefinitely. It could also be a difference in CI runner, but the memory notification
process isn't normally needed (only to help diagnose memory exhaustion errors).