Docker: MXE: reduce number of disused packages installed.
The CI runner has a hard limit of 180 minutes, and installing *all* MXE packages easily exceeds this time.

This commit also fixes a portability issue in the DCMA Wt webserver, but Wt is not yet available for cross compilation
due to lack of knowledge about which libraries to link. It also takes a long time to download all the Wt dependencies.