Docker: split MXE build into two parts.
The MXE cross-compile was taking too long to integrate into the existing CI process, so the MXE Docker image was split
into two parts: one for the toolchain and dependency preparation, and one for the DICOMautomaton compile. The toolchain
still takes too long to build from scratch, so I'll manually re-build the image as needed and push it to Docker Hub. The
image can be pulled for DICOMautomaton compiling during CI -- hopefully bandwidth won't be problematic (the MXE
toolchain is currently ~6GB).

Note that Ygor, Ygor Clustering, and Explicator are built twice, once during the toolchain construction (to verify the
toolchain successfully builds libraries and executables) and once as part of the DICOMautomaton builds (to make the MXE
toolchain container last as long as possible without having to re-build from scratch).