Added Repeat operation as first test of operation nesting.
The internal interface was simplified, relying on transitive access of children through each OperationArgPkg (so now
each OperationArgPkg knows how many children it has, and no special parameters are required access them). Children of an
OperationArgPkg are now properly encapsulated as private members of the OperationArgPkg class.

Note that a given OperationArgPkg still does not know to invoke itself; it has to defer to the Operation Dispatcher.
While this design is a bit tedious, it helps protect calling inappropriate operations (e.g., in the webserver).
Actually, the whitelist is only superficial at the moment, but at least the whitelist could be hoisted out as a
replacement / tailored dispatcher.