Cleaned up clang-format.
Used `whatstyle` to refine the clang-format style by minimizing the total changes to the code base. There are still a
few annoyances that stop me from running on the entire code base, namely missing brackets-and-spaces rules and handling
of comments. I probably need to pepper long comments that contain diagrams or delicate line breaks with 'clang-format
off' and 'clang-format on', but I haven't yet. It will be nice to have a consistent style that I can use to autoformat
while refactoring, but it just isn't important enough to justify at the moment.