EQDConvert: support both BED and EQDx conversions.
BED conversion were added to the EQDConvert routine. Various files and classes were renamed to reflect that the
underlying models are based on BED models, especially where EQDx is merely a ratio of BEDs. Some parameters were also
renamed to attempt to simplify the interface somewhat. Some parameter notes and descriptions were altered to make their
use more clear (especially differentiating EQDx and BED conversions). Integration tests were created to validate the
outputs; assumptions are noted in the tests.

While metadata is written to converted files, this commit does NOT enforce proper BED handling for any outputs. For
example, BEDs with different alpha/beta ratios cannot be summed. While it is technically possible to enforce this, it
would be trivial to circumvent and the distinction should be apparent in typical research workflows. A notice was added
to the operation description.