WIP: SFML_Viewer: preliminary dialog replacement.
This commit adds a simple, WIP dialog- or zenity-like replacement that can be use to simplify cross-platform text input.
This commit is incomplete, and has demonstrated that the current default fonts have inadequate glyph support, even for
core ASCII characters. It is also still unclear if a new dialog box is even required. It might make more sense to embed
a text interface inside the standard window, even if only normally visible for user input and advanced functionality or
debugging.