SFML_Viewer: integrate SFML-based dialog replacement.
I attempted to use a separate thread to manage the input dialog window and events, but it turns out that neither SFML
nor X11 seem to (easily) support multi-threaded window management. For example, creating a separate SFML render window
in a separate thread with a separate event queue caused dozens of data races to be caught in ThreadSanitizer. After a
lot of painful debugging and many whole machine lock-ups, I switched back to a single-threaded approach similar to the
existing plotwindow approach. Even this caused intermittent lock-ups in a VM where 3D acceleration passthrough was
enabled. Performance will correspondingly be limited.