Replace contours_with_meta with upstream contour_collection<> class.
This commit merged the Segmentation_History data member into metadata, leaving contours_with_meta as a simple alias to
contour_collection, so it made sense to remove contours_with_meta to simplify contour access (slightly). This change
undoes what is possibly the oldest bit of code in the DICOMautomaton code base, which was originally written
approximately 12 years ago. A fair amount of old-style member data selectors were also removed, and the
Segmentation_History machinery was degraded significantly. While it all still worked, the functionality was no longer
being used and reduced overall flexibility in other, more necessary parts of the code base.

The SegmentationHistory metadata will eventually be revamped or removed altogether; at the moment it isn't honoured in
enough places to ensure it is consistent, and it might be better to design around the problem and develop more general
spatial selectors that don't depend on tracking history. TBD.