WIP: added SDL-based interactive driver program.
The aim of this initiative is to provide a more standard GUI program to support interactive use-cases. Life the SFML
viewer, it will also be able to be launched as an operation (i.e., midway through some automated workflow), but will
provide a much amount of functionality compared with the SFML viewer -- hopefully as much as the webserver and SFML
viewer combined.

While the use of ImGui could have been added to the SFML viewer, it seems best from a portability angle to create a new
operation. Much of the SFML code used hard-coded details that I want to get away from, such as the position of
Image_Array data within the main viewer window.

This commit adds a minimal, useless SDL-based viewer operation, but it is not registered with Operation_Dispatcher(), so
is not yet accessible.