CI: provide alsa lib (libasound2) for AppImage testing.
The AppImage packager omits libasound2 and requires the system to provide their own copy. SDL2 requires alsa-lib, but
CMake doesn't know about it since it's only required at runtime. This commit adds a copy of alsa-lib to all
test systems that test with the AppImage.

Note that an alternative, simpler approach would be to install SDL2 on the test system. This would ensure all system
libraries required by SDL2 are available. For the time being I'll try keep the test systems lean.