SDL_Viewer: enable optional array viewing.
The previous implementation, which was inherited from the SFML_Viewer, requires that a valid Image_Array is available.
It is based on iterators. Default-constructed iterators not attached to a valid object are undefined behaviour, so
iterators are not well suited for this. In order to support optional viewing of images in a reasonable way, I replaced
the 'cached iterator' approach with iterators generated on-demand, as needed. It is slower and might still require some
caching, but at least supports optional image viewing.