Add GNU GSL lib to auxiliary program.
This lib link should have been added earlier when the GNU GSL code was being made optional, but normally the symbols are
resolved when linking with libygor, which itself includes GNU GSL. However, when compiling static libraries or with GSL
disabled a linker error previously could occur. A better solution would be for Ygor to provide a proper CMakeLists
Find script, but this is a reasonable solution since DCMA is likely to have GNU GSL enabled/disabled whenever Ygor does
anyways.