Nix: refactor nixpkgs, config, and overlays.
The new layout seems more idiomatic, and is cleaner in the sense that it effectively adds local ygor, ygorclustering,
explicator, and dicomautonaton derivations to nixpkgs, so treats them as 'upstream' which opens them up to be treated as
just another set of packages. I'm still not sure to what extent I can modify stdenv and implement cross-compilation with
the new layout, and note that default.nix was fully removed because I'm not sure what to put in it at the moment.