ConvertMeshesToPoints: add random surface sampling.
This commit adds support from upstream (Ygor) for randomly sampling the surface of an arbitrary surface mesh.
It is surprisingly fast, at least for meshes with several thousand faces and point clouds with several tens of thousands
of sample points. Loading an STL mesh file, extracting a point cloud with 1mm sampling density, and writing the point
cloud to file were effectively instantanious. It might be feasible to perform a point cloud 'thinning' operation
afterward to improve the point cloud quality.