Test that smesh exports can be loaded as smeshes.
This simple check is good enough to detect export failures, but will generally not catch subtle bugs. The integreity or
validity of the file is also not checked, except that the file successfully parses as a surface mesh. A more in-depth
test tool is required to perform more in-depth testing.