Added operation to make a transformation object from text.
The existing primitive types all map to Affine transformations, but the interface could support others. This operation
will likely replace the other operations that accept ad-hoc transformation strings. This will help to reduce
duplication, and also make it easier to enhance transformations.

While the operation essentially re-used one of the ad-hoc interfaces, it also added chained transformations, and
spatial mirroring. The transformations are also deferred, and can thus be written to/from file, tagged with metadata,
and applied to all spatial objects that can be transformed.