Separate transform creation from application.
This commit completes the separation between transformation creation and application to all major spatial object types.
Now, all the old 'transformation' application operations require a transformation object to have been created
previously. This simplifies development of new transformation types and implements DRY for a user-facing parser that was
problematic.

Additionally, all the 'TransformXYZ' operations were renamed to 'WarpXYZ' to (hopefully) more clearly reflect that the
transformation is spatial. The earlier names were hopelessly generic and ambiguous, so I think this is an improvement.
One downside it that it's less obvious, but that's a problem for documentation.