Add a non-trivial example (double-threaded screw).
This example highlights a non-trivial mesh processing operation. It performs a complicated boolean operation on a
non-manifold mesh via the 2D slicing method.

I'm not yet sure how or where to best include examples. I want to run them as part of CI, but I also want to make them
accessible and not obfuscate any details. For now, I'll include them in the integration tests. When a critical mass is
achieved, I'll consider reworking them.