Support fuzzy operation names and explicit aliases.
Often, a given operation could reasonably have several names that communicate what it can do. For example, the
HighlightROIs operation can be understood as an operation that 'in-paints' or 'out-paints' the voxels/pixels of an
image, but it can also be viewed as a way of converting bounding information from a contour representation to an image
mask. Both names are valid in a given context, but it might be difficult to search for HighlightROIs when one is looking
for a conversion operation. Operation name aliases address this problem. They also make it easier to make lightweight
name changes that are transparent to end-users.

Support for aliases uses the Explicator machinery, which also provides fuzzy matching. Finally, after many, many years,
DCMA is tolerant of minor stylistic changes in operation names, like capitalization, use of underscores, and
pluralization. With fuzzy matching comes risk of selecting the wrong operation, so a warning is emitted everytime a
non-exact match is made. Also, if the input is too distinct from all lexicon entries (composed of the operation names
and aliases) then it is explicitly rejected.

Integration tests were added to exercise fuzzy matching and alias functionality, and a basic form of documentation was
added to list all aliases alongside canonical names.