Expose Drover partitioning backend.
I was planning to use this to implement something like the GroupImages operation, but decided against it for now. In the
meantime, exposing the side-effect-free Partition_Drover backend opens up a few more opportunitites for more robust
grouping. The generic metadata extractor operation will also be useful.

This switch also brings stability to the ForEachDistinct operation, which previously would re-order partitions using the
key-value list. I replaced the std::map that caused this issue in favour of a std::list and an external index. I also
added an alias to the GroupImages operation to make it slightly more discoverable.