Simple_Meshing: accept disoriented contours.
Previously, contours with inconsistent orientations that were ent to the amalgamation routine caused meshing to fail.
While this behaviour is valid, it leaves a lot to be desired when accepting contours from real data sources, which often
ignore contour orientation.

This commit permits contours with non-uniform orientations to be handled as though they all have the same orientation.
Note this this is a lossy assumption, but may be more useful for the typical case.

Ideally, a more sophisticated method will replace the current amalgamation script (i.e., 2D boolean with inner and outer
hole support), but I have no specific plans to do so at the moment.