SDL_Viewer: support multiple file selections.
This commit attempts to mitigate some of the pain of not having a first-class file picker widget. It allows selecting
and loading multiple files at once, though they all need to be in the same directory. It also allows specifying the root
directory explicitly, which is needed for Windows systems.