CI: shrink executables for patchelf
When the ImGui and ImPlot demos are compiled into dicomautomaton_dispatcher and dicomautomaton_webserver, patchelf fails
with:

    Setting rpath in ELF file ./AppDir/usr/bin/dicomautomaton_dispatcher to $ORIGIN/../lib
    ERROR: Call to patchelf failed:
    warning: working around a Linux kernel bug by creating a hole of 8593408 bytes in ‘./AppDir/usr/bin/dicomautomaton_dispatcher’
    maximum file size exceeded
    ERROR: Failed to set rpath in ELF file: ./AppDir/usr/bin/dicomautomaton_dispatcher

This causes AppImages to fail if the system is unable to provide needed libraries. Reducing the size of the executable
addresses the issue, though problems may persist in the future if the binary keeps increasing in size. A stop-gap would
be to aggressively strip binaries, optimize for space, or reduce dependencies.