SDL_Viewer: provide contouring fallback method.
The marching cubes contour extraction method is only available when CGAL support is available due to the polygon slicing
functionality. When this is not available, a simpler binary method is used. While the contours are jagged, realtime
refresh rates are possible with much larger images compareed with the marching cubes -> 3D mesh -> polyline extraction.