ContourViaThreshold: make marching-squares handle boundaries
This commit offsets and enlarges the marching squares grid so that the image boundary is fully enclosed by (virtual)
exterior voxels. This ensures all contours will properly close such that stray edges don't skip the image corners. With
limited testing, marching-squares seems to produce the exact same outputs as marching-cubes (excet with superior image
boundary handling) at a fraction of the computational effort. It also does NOT depend on CGAL, so should be available on
all platforms.