HighlightROIs: add WIP marching-squares-inverse algorithm
This commit adds a working implementation of an algorithm that will convert contours to a "mask" such that when
marching-squares is run on the mask, the original contours will (mostly) be reproduced. This is in contrast to a binary
mask, where the contours can wander within half of the largest voxel diagonal.

In many cases a binary mask is sufficient, but in some cases contour reproduction is critical. For example, dilation and
erosion morphological operations are much simpler when working on voxel grids, but if the conversion between contours
and voxel grids is lossy than small erosions/dilations can be extremely inaccurate.

Note that this implementation is extremely slow, and would benefit tremendously from spatial indexing.