HighlightROIs: expose receding-squares method
This implementation is still extremely slow. It would benefit tremendously from bounding checks or spatial indexing.
However, it is functional and may be useful without optimization.

Note that there might be edge cases not handled by various combinations of parameters. This implementation was developed
with the expectations that: (1) the entire image will be overwritten, (2) contours are accurate and selective, so that
ContourOverlap should be either honour_opposite_orientations or overlapping_contours_cancel, and that (3) the contour
detail and image grid resolution are sufficiently matched that it is uncommon for multiple contours to pass between
adjacent voxels.

Expectation (3) could significantly impact round-trip contour accuracy, so consider using high-resolution images and, if
possible, avoid pathological contours (e.g., multiple colinear contours separated by small distances).