HighlightROIs: rely on modified upstream Mutate_Voxels() for contour overlap
This commit also differentiates contours in volumetric cases, which previously tested overlap of all contours on all
images.

This commit also improves the way the 'ignore' option is computed, handling separately the vertical and horizontal
neighbours where possible. The overall method remains a simplistic one-pass approach that only considers only the
immediate left and immediate up neighbours to determine how the intensity of the (*this) voxel. Slight differences can
arise do to averaging and numerical imprecision, especially if the input comes from a marching algorithm which can
result in geometrical alignment and degenerate voxel inclusions.

Documentation of the various contour overlap options was also improved.