SDL_Viewer: support gapless contouring.
When dragging the mouse quickly relative to the frame rate, the immediate-mode approach can result in gaps and stutter
in the image mask. This commit switches to a line-segment approach where the current and previous mouse position are
considered to be joined with a straight line, and the mouse 'point' becomes a line-segment.

This approach also permits straight lines to be drawn (by holding shift) and row- and column-projected lines (shift +
control) relatively easily.

The current implementation is fairly slow. The full image is evaluated every time, regardless of where the mouse is.
Though slower, it avoids stutter and gaps, so seems worthwhile.