SDL_Viewer: support 3D contour brushes.
This commit adds a contouring image for every display image in the display image array. The geometry is matched.
Contours are only extracted for the current frame to avoid recomputation that the user won't see. This implementation
functions well, but is extremely slow. Especially 3D brushes. A faster approach is needed to accelerate contour mask
blitting, which currently runs on the main thread and potentially requires traversal of all voxels throughout the 3D
volume.