CellularAutomata: added simplistic 'gravity' mode for voxel values.
Voxel values are treated as densities. More dense voxels 'sink' to the bottom, while lighter values 'float' to the top.
The orientation of gravity can be altered. Note that due to conservation of mass, this effect results in obvious
'bubbles' that appear strange; this is analogous to negative gravity operating on low-density voxels while positive
gravity acts on high-density voxels.