Docker: avoid unnecessary costly rebuilds (Debian stable).
This commit adds logic that will only bother compiling custom dependencies (Ygor, Explicator, Ygor Clustering) when
upstream has diverged from the existing repo. The rationale is that the prior build was already compiled and installed,
so in cases where nothing has changed, recompilation should be avoided.

Note that this may cause issues when system updates are implemented, but in these cases the entire chroot cache can be
regenerated. Currently this requires manual intervention, so further tweaking will be required eventually.