SDL_Viewer: make auto-scaler aware of multi-channel images.
This commit forces the auto-scaler to disregard baked-in metadata centre and width directives, which may not correctly
handle multi-channel images. Per-channel metadata would fix this, but is not implemented or supported in any operations,
so relying on the SDL_Viewer autoscaler, where it is most likely to matter, seems like a reasonable compromise.