SDL_Viewer: roughed-in worker offload thread.
This commit has 'roughed-in' support, however there appears to still be some unprotected state or lack of
synchronization, so the worker thread is not used. The contention seems to be between the worker thread and the contour
pre-processor threads. I'm not happy with the state of the contour preprocessor, and think I can likely improve the
situation somewhat.

However, it's clear that a lot of work will be needed to successfully offload work on the main Drover class, likely even
tagging operations as read-only or write access to permit concurrent access. That could be a lot of work!