SDL_Viewer: compartmentalize sub-operations.
This commit makes use of the previously roughed-in worker thread to offload user operations on the central Drover
object. The previous issues with data races and synchronization were (most likely?) due to the arguments being passed to
the tuple_cat() function (done for convenience) being invoked in an unspecified order. One of the arguments was a mutex
lock, so being called out-of-order was caused frequent data race conditions and memory corruption. (Note that this
functionality was disabled in the previous commit specifically to avoid risk of corruption!)

Compartmentalization seems to have also helped in figuring out the dependencies between sub-operations. Opportunities to
offload more operations are being considered for future work.