SDL_Viewer: used timed mutex to skip locked data access.
This commit lets the script worker thread to block while non-blocked gui elements continue to be updated. This is
accomplished using a timed mutex with a small try-lock-for duration (a few microseconds) that normally should be
sufficient to lock but will skip the element if the mutex is held.

There is an annoyance in that the window settings (position, size) revert to the default. I'm not sure if this can be
fixed...