File_Loader: recursively find files when directories provided.
This commit also adds the distinction between implicitly-provided files (specified recursively within a directory) and
explicitly-provided files. File extension filtering is performed on implicit files, since it is often the case that
directories contain a mix of files. However, file extension filtering is NOT performed on explicit files, since the
extension can sometimes be misleading or irrelevant.

Warning messages are provided to the user when extension filtering is being used to ignore files, and when it could be
used to ignore files that have an unrecognized extension.