Dispatcher: added workaround for AppImageKit chdir() call.
This workaround is currently necessary because '/usr' is replaced by '././' in binaries, so a chdir() is needed to make
'./' point to the root of the AppDir. Unfortunately, linuxdeployqt, which is an alternative, is not available on all required platforms.

Note that another workaround that works prior to this commit is to provide absolute path names.

See the following issue threads for additional context:
- https://github.com/AppImage/AppImageKit/issues/172
- https://github.com/AppImage/AppImageKit/issues/989
- https://docs.appimage.org/packaging-guide/environment-variables.html#type-2-appimage-runtime
- https://github.com/AppImage/AppImageKit/issues/1035
- https://github.com/linuxdeploy/linuxdeploy/issues/123
- https://github.com/AppImage/AppImageKit/issues/508