GroupImages: provide automated partition key selection.
This feature is meant to simplify tasks like identifying which metadata key to use to create ~equal-size partitions.
This is a common problem in medical image analysis where there is a longitudinal, non-spatial dimension (e.g., epoch or
shuttle number for perfusion, diffusion b-value for diffusion-weighted MR imaging, and time for longitudinal
radiotherapy dose analysis).