Operations: separate Drover updates from return/exit.
Currently control flow meta-operations rely on throwing exceptions to determine truthiness. However, some operations
should be able to update the Drover instance before reporting a failure. For example, the internals of a while loop,
which should continuously mutate state on each loop, rather than propagating the changes only at the end and only if the
loop terminates successfully.

I'm still trying to figure out the specifics of how throwing and false-truthiness should be interpretted. At the moment,
they are considered the same thing but allow for greater freedom (i.e., returning non-error false truthiness, and
throwing to quickly leave a scope with an error message).