ModifyImageMetadata: support macro expansion and time extraction.
This commit adds support for metadata key-value-based macro expansions. It also adds support for extracting the number
of seconds represented by a time, date, or date time in typical formats. Upstream support was added to Ygor to support
extracting fractional seconds when the time part has trailing decimals, e.g., 'HH-MM-SS{.SSSSS}'.

This commit is not specific to images and should be hoisted out of ModifyImageMetadata and into a more widely-applicable
operation. I plan to do this in the future.

Also note that use of YgorTime was only used because the necessary add-ons to std::chrono are only available in C++20.
This is also likely to (eventually) be updated.