SDL_Viewer: contouring: handle case with multiple overlapping images.
This use-case isn't particularly well supported, since it's annoying to figure out whether images are part of a
volumetric set or not. However, for repeated image sets (e.g., perfusion, diffusion, 4DCT, anything with a temporal
component) this commit provides a way to generate and save contours and apply morphological operations at the expense of
additional computational effort.

I'm not happy wth the current state of contouring, so will likely increase the amount of precomputation or offload to
another thread.