SDL_Viewer: proactively set LIBGL_ALWAYS_SOFTWARE=1 within AppImages.
This might be overkill for generic AppImage deployments, but will help increase portability. The downside, reduced
OpenGL performance, can be overridden by setting LIBGL_ALWAYS_SOFTWARE=0. (If this is supported by the host system, then
disabling software emulation should generally be safe.)

Note that this workaround is addresses the error 'Unable to create an OpenGL context for SDL: Could not create GL
context: GLXBadFBConfig' which is encountered when hardware support for recent OpenGL is not present (e.g., when running
OpenGL 3.0 AppImages inside Virtualbox).