Cross-compile: more tinkering to get armv8 to cross-compile with zig-cc or musl toolchain.
This script adds a few notes and workarounds currently needed to cross-compile. For some reason there are missing time
and thread routines when compiling for armv8. I was able to successfully cross-compile DMCA for armv8, but the resulting
binaries were not actually static. They still rely on the musl linker being found on the host computer. Confirmed they
did not work on an armv8 phone.

I think a more robust alternative will be to use an established and maintained toolchain, such as Alpine Linux.