Fallback to an identity conversion if iconv does not support required conversion.
This commit permits iconv_open() to fallback on a utf8-utf8 conversion if the requested conversions are not available.
This is needed on Alpine since musl provides a fraction of the conversions that glibc's iconv does. However, I suspect
that in many cases this should be OK since the majority of data encountered these days *should* be utf8, and at the
moment all musl builds assume Linux, which is utf8-native.