SDL_Viewer: encapsulate mesh loading and rendering in a class.
This change permits meshes to be loaded once and continually re-displayed without having to reload the vertex data.
It also helps to hide the details of registering with OpenGL, which were fairly distracting.

Note that this commit removes the frame_count-based rotation -- this functionality will (eventually) be re-added as a
viewport effect, though no ETA is available.