WIP: SDL_Viewer: embed raw opengl mesh in imgui window.
This currently does not work, and I have no idea what the problem is. For sanity purposes, as well as having the largest
possible view of the 3D scene, I'll revert this commit and draw the opengl scene in the back, before an imgui calls each
frame. Committing this just in case I want to jump back and try something out in the future.