SDL_Viewer: maintain aspect ratio in model space.
This commit enables aspect ratio normalization in model space. Meshes should now render square, regeardless of the
user's window dimensions (GL viewport) or the aspect ratio of the model (GL model space). The downside is that meshes
can be rendered in a small area relative to the total available screen space.

I hope to give more flexibility in the perspective and scale of rendered meshes in a future commit.