De-CGALify marching cubes implementation.
This commit decouples the DCMA marching cubes implementation from CGAL, ensuring it is available regardless of whether
CGAL can be compiled. Note that CGAL is currently disabled in some CI builds to either reduce memory consumption, or
because it does not support the particular system.

Note that the internal marching cubes implementation, as alluded to in recent commits, may be moved to Ygor.