Support reading and visualizing deformation fields.
This commit adds some missing tags and connects the DICOM file REG modality reader with the deformation_field class to
actually load the vector grid data.

Loading is currently much slower than it needs to be, and visualization is rudimentary. I plan to address both of these
issues in the near future.