Avoid std::filesystem::absolute on Windows.
Use of absolute() causes UNC paths on Windows to become invalid such that std::filesystem::exists() does not recognize
the resulting path, and std::fstream() can't access the file. This commit makes absolute() an optional step, only
retaining the transformation if the resulting file can be reached.

This commit also replaces many uses of std::string filenames with std::filesystem::path, which is both more robust, and
more clearly signals intent. However, std::filesystem::path has a string operator() that makes transition between
string and fs::path relatively transparent. So there are almost certainly cases where paths are converted back to
strings.

Note for future reference: UNC paths on Windows need to be converted from
    \\server_name\share_name\path\to\dicomautomaton_dispatcher.exe
to
    \\\\server_name\\share_name\\path\\to\\dicomautomaton_dispatcher.exe
(This seems a bit ridiculous and is a huge impediment to cross-platform computing. I don't understand how this is the
current state of affairs??)