SDL_Viewer: simplify file loading.
This commit trades the fallback (SDL-based) file loader for a simplified file loading process + a new script loader.
I think it would potentially be useful to have an SDL-based file selector dialog as a fallback, but it doesn't seem to
be needed at the moment. If needed, I think I would prefer to launch a second instance/window or even use a separate
process (like use of zenity by PFD).

This commit also adds the ability to imbue the file selection dialog with custom user data. It isn't currently used (was
added for a different design that was reverted) but I think it's a useful feature to keep around.