Replace large macro with generated file.
The macro approach to imbuing standard scripts was causing issues with CMake, leading to sporadic parsing failures due
to memory limits being exceeded for CMake strings. This commit replaces the CMake-based embedding process with a script
(that can be run idempotently) that is more robust to large string memory usage.

Note that the current approach might still lead to command line length limits and memory usage being exceeded. It might
also fail if the hexdump program is not available.