Add QueryUserInteractively operation.
This commit is rather larger since I had to alter the main Operation_Dispatcher process to support macro replacement,
add support for value re-writing in the OperationArgPkg struct, alter SDL_Viewer script handling, change the signature
of the InvocationMetadata parameter in all function calls, and of course build the various new operations and
functionality.

The (global) parameters are somewhat protected from accidental use by the caller via having to prepend a '$' for all
expansions, though I'm not fully sure this is a collision-less approach. I know some operations/methods already use '$'
for expansions, but I can't tell if also being able to expand globals in the same way is a positive or negative.

Everything seems to work reasonably well, and tests pass, but tweaking and further work may be needed.