Link directly to GitHub.
Relative links work in GitHub, but not GitLab. To avoid messing around trying to figure this out, I'll default to GitHub
for now.

(Is there a way I can define a variable in Markdown sources? This would let me refer to GitHub on GitHub, GitLab on
GitLab, and locally if needed.)