SDL_Viewer: add toggleable face smoothing.
This essentially controls the shading model, with either interpolates the fragment colour between vertices, or uses the
one of the vertices (the 'provoking' vertex) for the entire face. Unsmoothed faces make it easier to see the geometry
for low-resolution models, whereas it can detract when the detail is irrelevant or would be distracting.