SDL_Viewer: avoid drawing scale bar when changing colour map.
This addresses an OpenGL error in which the draw list submitted by ImGui referenced a non-existent texture. It manifest
as the message '(line 1329) : Unknown error(1282)'.

There may still be an issue with the display image since it can also be replaced by a user action while the draw list is
active. I can't produce an error however. One way to fix would be to defer replacing the image until the end of the
frame. Or, as was done here for the scale bar image, postpone drawing the image until after the swap has occured.