OrderImages: support stable image array sorting.
This commit adds support for 'external' (i.e., inter-array) image sorting. This compliments the
'internal' (i.e., intra-array) image sorting that was previously available.

Image selectors make uncommon to need to rely on image array sorting, however it is useful when presenting images in a
standard format. For example, with increasing SeriesNumber.