SDL_Viewer: implement manual image feature selector.
This can be used to extract features from two image arrays with different FrameOfReferenceUID. The selector is
functional and can export either point sets or a rigid transformation between two point sets. However, more experience
is needed with the current interface in order to figure out how to make it ergonomic and more useful.